r83 | ptpb | 2005-04-11 10:23:53 +0200 (Mo, 11. Apr 2005) | 10 Zeilen Geänderte Pfade:    M /trunk/src/paw_waves1.f90    M /trunk/src/paw_waves2.f90
The individual operations of waves etot have been packaged, so that
the routine becomes more transparent and so that the parallelization
is mostly concentrated in the intermediate layer. This is necessary
for the k-point parallelization.  Kaestner's Conjugate gradient has
been disabled. A call of this option will provide information on how
to repair it.
----------------------------------------------------------------------
paw_waves1.f90 paw_waves2.f90 CVS:
----------------------------------------------------------------------